Include mode and mountpoint in config file
Closes #14.
