+++ rmfuse/config.py
@@ -1,7 +1,8 @@
 # Copyright 2021 Robert Schroll
 # This file is part of RMfuse and is distributed under the MIT license.
 
-import configparser
+from configparser import ConfigParser
+import enum
 import inspect
 import logging
 import sys
@@ -13,20 +14,34 @@ from rmrl import render
 log = logging.getLogger(__name__)
 CONFIG_FILE = xdg_config_home() / 'rmfuse' / 'config.ini'
 
+class FSMode(enum.Enum):
+    meta = 'meta'
+    raw = 'raw'
+    orig = 'orig'
+    annot = 'annot'
+
+    def __str__(self):
+        return self.name
+
 def _get_render_defaults():
     defaults = inspect.getfullargspec(render).kwonlydefaults
     del defaults['progress_cb']
     return defaults
 
 DEFAULTS = {
+    'mount': {
+        'mountpoint': '',
+        'mode': FSMode.annot
+    },
     'render': _get_render_defaults()
 }
 
 LOOKUP_FUNCS = {
-    bool:  'getboolean',
-    float: 'getfloat',
-    int:   'getint',
-    str:   'get'
+    bool:   lambda config, key: config.getboolean(key),
+    float:  lambda config, key: config.getfloat(key),
+    int:    lambda config, key: config.getint(key),
+    str:    lambda config, key: config.get(key),
+    FSMode: lambda config, key: FSMode(config.get(key))
 }
 
 def to_dict(config, defaults):
@@ -36,14 +51,14 @@ def to_dict(config, defaults):
             res[k] = to_dict(config[k], defaults[k])
         else:
             try:
-                res[k] = getattr(config, LOOKUP_FUNCS[type(defaults[k])])(k)
+                res[k] = LOOKUP_FUNCS[type(defaults[k])](config, k)
             except (TypeError, ValueError, KeyError):
                 log.warning(f'Failed to read "{k}: {config[k]}" as {type(defaults[k])}')
                 res[k] = defaults[k]
     return res
 
 def get_config(section=None):
-    config = configparser.ConfigParser()
+    config = ConfigParser()
     config.read_dict(DEFAULTS)
     if CONFIG_FILE.is_file():
         config.read_file(CONFIG_FILE.open('r'))
@@ -52,7 +67,7 @@ def get_config(section=None):
     return to_dict(config, DEFAULTS)
 
 def write_default_config():
-    config = configparser.ConfigParser()
+    config = ConfigParser()
     config.read_dict(DEFAULTS)
     if CONFIG_FILE.exists():
         log.warning(f'{CONFIG_FILE} already exists; output to stdout intead')
